Multiplayer!!!
This was not fun. That's about it. I don't really understand how the code works anyways so I'm not going to try to explain what I did. basically a "gameObject" is passed to each client every 120th of a second and the clients send back an action that tells if they have moved.